requirements.txt
Django==5.0.4
dj_database_url==2.1.0
whitenoise==6.6.0
celery==5.4.0
django-environ==0.11.2
gunicorn==22.0.0
djangorestframework==3.15.1
psycopg2==2.9.9
black==24.4.0
django-admin-interface==0.28.6
django-admin-autocomplete-filter==0.7.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
django-anymail==10.3
jsonpatch==1.33
django-allauth==0.62.1
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
pika>=1.3.2